Membuat Fitur CRUD, Trash, Restore, Force Delete, Find by Status, Find by Category, Comment Artikel

+++ routes/api.php
@@ -2,11 +2,10 @@
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
-use App\Http\Controllers\Api\User\ProfileController;
-use App\Http\Controllers\Api\Content\ContentController;
 use App\Http\Controllers\Api\Content\ArticleController;
-use App\Http\Controllers\Api\Authentication\AuthenticationController;
+use App\Http\Controllers\Api\Content\CommentController;
 use App\Http\Controllers\Api\Content\CategoryController;
+use App\Http\Controllers\Api\Authentication\AuthenticationController;
 
 /*
 |--------------------------------------------------------------------------
@@ -32,23 +31,33 @@ Route::post('/login', [AuthenticationController::class, 'login']);
 Route::group(['middleware' => ['AuthBasicApi']], function () {
     Route::get('/logout', [AuthenticationController::class, 'logout']);
 
-    //test route category
+    //Category Route
     Route::post('/category', [CategoryController::class, 'store']);
     Route::get('/category', [CategoryController::class, 'index']);
     Route::get('/category/{id}', [CategoryController::class, 'show']);
     Route::put('/category/{id}', [CategoryController::class, 'update']);
     Route::delete('/category/{id}', [CategoryController::class, 'destroy']);
+
+    //Article Route
+    Route::post('/article', [ArticleController::class, 'store']); //belum bisa upload file
+    Route::put('/article/{id}', [ArticleController::class, 'update'])->middleware('Author'); //pengerjaan
+    Route::delete('/article/{id}', [ArticleController::class, 'destroy'])->middleware('Author');
+    Route::get('/article/search/{keyword}', [ArticleController::class, 'search']);
+    Route::get('/article/user/{id}', [ArticleController::class, 'showByUser']);
+    Route::get('/article/status/{id}', [ArticleController::class, 'showByStatus']);
+    Route::get('/article/category/{id}', [ArticleController::class, 'showByCategory']);
+    Route::get('/article/trash/{id}', [ArticleController::class, 'showTrash']);
+    Route::get('/article/trash', [ArticleController::class, 'showTrashAll']);
+    Route::get('/article/restore/{id}', [ArticleController::class, 'restore']);
+    Route::get('/article/force/{id}', [ArticleController::class, 'forceDelete']);
 });
 
 //User Panel
-Route::get('/content', [ContentController::class, 'index']);
-Route::get('/articles', [ArticlesController::class, 'index']);
-
-//test route articles
-Route::post('/article', [ArticleController::class, 'store']); //pengerjaan
 Route::get('/article', [ArticleController::class, 'index']);
 Route::get('/article/{id}', [ArticleController::class, 'show']);
-Route::put('/article/{id}', [ArticleController::class, 'update']); //pengerjaan
-Route::delete('/article/{id}', [ArticleController::class, 'destroy']);
-Route::get('/article/search/{keyword}', [ArticleController::class, 'search']); //pengerjaan
-Route::get('/article/category/{id}', [ArticleController::class, 'showByCategory']);//pengerjaan 
+
+//test route commert
+Route::post('/comment', [CommentController::class, 'store']);
+Route::get('/comment', [CommentController::class, 'index']);
+Route::delete('/comment/{id}', [CommentController::class, 'destroy']);
+Route::get('/comment/force/{id}', [CommentController::class, 'forceDelete']);
